US100 Log hooks added to the event channel
