+++ Assets/Scripts/Logging/Runtime/IEventLogger.cs
@@ -1,3 +1,4 @@
+using UnityEngine;
 
 namespace Logging
 {
@@ -43,6 +44,25 @@ namespace Logging
         Network = 1 << 5,
         // All categories enabled.
         All = ~0
+    }
+
+    /// <summary>
+    /// Represents a single log entry transmitted via EventChannel.
+    /// value-based type for safe broadcast.
+    /// </summary>
+    public class LogEvent
+    {
+        public LogLevel Level { get; }
+        public LogCategory Category { get; }
+        public string Message { get; }
+        public Object Context { get; }
+        public LogEvent(LogLevel level, LogCategory category, string message, Object context = null)
+        {
+            Level = level;
+            Category = category;
+            Message = message;
+            Context = context;
+        }
     }
 
     // Event Logger is implemented using Unity's Debug class.
