--- Assets/Scripts/Logging/Runtime/LogEvent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1c4de7b3e475f0c44897ec5580491261
