+++ Assets/Scripts/Logging/Runtime/LogEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Logging
+{
+    [UnityEngine.CreateAssetMenu(fileName = "LogEventChannel", menuName = "PsycheOpoly/Events/Log Event Channel")]
+    public class LogEventChannel : EventChannel<LogEvent> { }
+}
