+++ Assets/Scripts/Logging/Runtime/LogEventChannel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cd25b0a1e7164b14c8534a66ac1b8ceb
