+++ Assets/Scripts/Logging/Runtime/LogHook.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dec83eef7c071474b8493bb2bd20739a
